move test files to suite folder
