+++ test/suite/extension.test.ts
