+++ test/suite/previewContainer.test.ts
@@ -1,9 +1,9 @@
 import * as assert from "assert";
 import * as vscode from "vscode";
-import {PreviewContainer} from "../src/previewContainer"
-import Reporter from "../src/telemetry";
-import vscodeUtils from "../src/vscodeUtilities";
-import {settings} from "../src/settings"
+import {PreviewContainer} from "../../src/previewContainer"
+import Reporter from "../../src/telemetry";
+import vscodeUtils from "../../src/vscodeUtilities";
+import {settings} from "../../src/settings"
 
 
 /**
