+++ test/suite/toAreplLogic.test.ts
@@ -1,9 +1,9 @@
 import * as assert from "assert";
 import * as vscode from "vscode";
-import {ToAREPLLogic} from '../src/toAREPLLogic'
-import Reporter from "../src/telemetry";
-import vscodeUtils from "../src/vscodeUtilities";
-import {settings} from "../src/settings"
+import {ToAREPLLogic} from '../../src/toAREPLLogic'
+import Reporter from "../../src/telemetry";
+import vscodeUtils from "../../src/vscodeUtilities";
+import {settings} from "../../src/settings"
 import { EOL } from "os";
